llm-tracker/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.7.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.3
redis==5.0.1
python-dateutil==2.8.2
pytz==2024.1